Add an initial implementation of a resolved-RIB.
Background -- we want the gRIBI server to be able build both a RIB that
is efficient for it to do gRIBI operations with, as well as a software
emulated 'resolved RIB' / 'FIB' whereby we can say something like "where
will this packet go based on what gRIBI is programmed?". This package
starts this implementation.

  * (M) aft/oc.go
    - Update generated code.
  * (M) go.mod
  * (M) go.sum
    - New dependencies on kentik/patricia
  * (A) ocrt/oc.go
  * (A) ocrt/update.sh
    - Generated routing specific OpenConfig Go structs.
  * (A) ocrt/yang/deviations.yang
    - Deviations to reduce the size of the generated oc.go.
  * (M) rib/rib.go
    - Fix handling of implicit replace.
  * (A) sysrib/sysrib.go
  * (A) sysrib/sysrib_test.go
    -  Initial system RIB implementation. This package currently parses
        OpenConfig to build a list of all of the connected routes that
        are on the system.