Handle rib-ack op responses received after a fib-ack (#183)
Previously, if a client in RIB_AND_FIB_ACK mode received a rib-ack response for
an op after receiving the fib-ack, the client would return a "op not found"
error. This happened because the client deleted the op from the pending op queue
while handing the fib-ack, and then the client would not find the op in the
pending queue when processing the following rib-ack

Normally, the client expects the rib-ack to precede the fib-ack, so this problem
only happens in the edge case where the fib-ack is received by the client before
the rib-ack.

As a workaround, the client will now return a generic result without op detail
to the rib-ack. The client can't fetch op details for the rib-ack because the op
will have already been deleted by the fib-ack response handling. In the future,
we can consider caching pending ops to ensure the details are available.

Co-authored-by: Nick Hawke <nhawke@google.com>